translated previous tests, moved one
